requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
flask==2.2.3
werkzeug==2.2.3
waitress==3.0.0